police scene 2 (desk)
(Completed Features)

Desk and Police Interaction:

Added a desk element to the game.
Implemented a police character eating donuts , the player must "bunch" the police to proceed.

Police Disappearance on Interaction:

 implemented the functionality for the police character to disappear after being "bunched" by the player.
Issues to Address:

(Issues)

PlayerStats Error on Police Interaction:
Identified an issue in the PlayerStats script related to player interaction with a police character eating donuts.
When the player is hit by the police character and subsequently touches hot coffee, a NullReferenceException occurs in PlayerStats.TakeDamage (line 46).

(Work in Progres)

Currently in the process of implementing the functionality to acquire new abilities after defeating a police character who is eating donuts.

(To-do)

Implement visible and invisible boxes in the game environment.
Design doors into the game environment.
Enable the enemy to take damage